Moving CA1820 over. This analyzer used to implement a base type called AbstractSyntaxNodeAnalyzer. I've dropped that type since no other analyzer uses it and just moved a couple of abstract method from that type to the base analyzer itself. The analyzer's message also has no fillins and so replaced the stub generated message.
