+++ PortFxCop/src/System.Runtime.Analyzers/CSharp/System.Runtime.CSharp.Analyzers.props
@@ -9,7 +9,7 @@
    <PropertyGroup>
       <CodeAnalysisRuleSetOverrides>
          $(CodeAnalysisRuleSetOverrides);
-         <!--INSERTRULESETOVERRIDES-->
+         -Microsoft.Performance#CA1820;
       </CodeAnalysisRuleSetOverrides>
    </PropertyGroup>
 </Project>
