+++ PortFxCop/src/System.Runtime.Analyzers/VisualBasic/System.Runtime.VisualBasic.Analyzers.props
@@ -8,8 +8,8 @@
   -->
    <PropertyGroup>
       <CodeAnalysisRuleSetOverrides>
-         $(CodeAnalysisRuleSetOverrides);
-         <!--INSERTRULESETOVERRIDES-->
+        $(CodeAnalysisRuleSetOverrides);
+        -Microsoft.Performance#CA1820;
       </CodeAnalysisRuleSetOverrides>
    </PropertyGroup>
 </Project>
